+++ src/assets/svgs/index.ts
@@ -1,1 +1,2 @@
-export {}
+export { ReactComponent as HomeIcon } from './home.svg'
+export { ReactComponent as ManageIcon } from './manage_accounts.svg'
