+++ src/components/Layout/layout.module.scss
@@ -7,12 +7,8 @@
   background-color: colors.$BACKGROUND;
 
   .mainContainer {
-    padding: sizes.$HEADER_HEIGHT 40px 60px;
+    padding: sizes.$HEADER_HEIGHT 40px 40px;
     margin-left: sizes.$LNB_WIDTH;
-  }
-
-  svg {
-    fill: currentcolor;
   }
 }
 
@@ -20,70 +16,62 @@
   position: fixed;
   top: 0;
   right: 0;
-  left: sizes.$LNB_WIDTH;
+  left: 0;
   z-index: levels.$HEADER;
-  width: calc(100% - sizes.$LNB_WIDTH);
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  height: sizes.$HEADER_HEIGHT;
   padding: 0 40px;
-  background-color: colors.$BACKGROUND;
+  background-color: colors.$WHITE;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 20%);
 
-  ul {
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    height: sizes.$HEADER_HEIGHT;
-    border-bottom: 1px solid colors.$GRAY50;
-
-    li + li {
-      padding-left: 30px;
-    }
+  h1 {
+    font-size: 22px;
+    font-weight: 700;
+    color: colors.$GRAY800;
+    text-transform: uppercase;
+    letter-spacing: 0.05em;
   }
 
-  .notification {
-    position: relative;
-
-    &::after {
-      position: absolute;
-      top: 0;
-      right: 0;
-      width: 5px;
-      height: 5px;
-      content: '';
-      background-color: colors.$SECONDARY_RED;
-      border-radius: 50%;
-    }
+  span {
+    margin-right: 20px;
+    color: colors.$GRAY800;
   }
+}
 
-  .profile {
-    display: flex;
-    align-items: center;
+.lnb {
+  position: fixed;
+  top: sizes.$HEADER_HEIGHT;
+  left: 0;
+  z-index: levels.$LNB;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  width: sizes.$LNB_WIDTH;
+  height: 100%;
+  padding: 40px;
+  background-color: colors.$SECONDARY;
+  box-shadow: 4px 0 10px 0 rgba(0, 0, 0, 4%);
 
-    &:hover {
-      span {
-        color: colors.$GRAY300;
+  .navigation {
+    a {
+      display: flex;
+      align-items: center;
+      padding: 20px 22px;
+      border-radius: 10px;
+      transition: all 0.1s ease-in-out;
+
+      svg {
+        margin-right: 12px;
+        fill: currentColor;
       }
-    }
 
-    img {
-      width: 40px;
-      height: 40px;
-    }
-
-    span {
-      margin-left: 10px;
-      font-size: 14px;
-      font-weight: 500;
-      line-height: 16px;
-      color: colors.$GRAY800;
-      transition: 0.1s ease-in-out;
-    }
-  }
-
-  svg {
-    fill: colors.$GRAY800;
-    transition: 0.1s ease-in-out;
-
-    &:hover {
-      fill: colors.$GRAY300;
+      &.isActive {
+        color: colors.$WHITE;
+        background-color: colors.$PRIMARY;
+      }
     }
   }
 }
