+++ src/components/index.ts
@@ -1,4 +1,3 @@
-
 import Layout from './Layout'
 
 export { Layout }
