+++ src/index.tsx
@@ -2,8 +2,6 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from 'react-router-dom'
 import { RecoilRoot } from 'recoil'
-import { QueryClient, QueryClientProvider } from 'react-query'
-import { ReactQueryDevtools } from 'react-query/devtools'
 
 import reportWebVitals from './reportWebVitals'
 import './styles/index.scss'
@@ -11,18 +9,14 @@ import './styles/global.scss'
 
 import Routes from './routes'
 
-const queryClient = new QueryClient({
-  defaultOptions: { queries: { refetchOnMount: false } },
-})
-
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 root.render(
   <React.StrictMode>
-      <RecoilRoot>
-        <BrowserRouter>
-          <Routes />
-        </BrowserRouter>
-      </RecoilRoot>
+    <RecoilRoot>
+      <BrowserRouter>
+        <Routes />
+      </BrowserRouter>
+    </RecoilRoot>
   </React.StrictMode>
 )
