+++ src/states/index.ts
@@ -1,1 +1,1 @@
-export { userDataState } from './userdataState'
+export { userDataState } from './userDataState'
