+++ src/styles/constants/_colors.scss
@@ -1,3 +1,12 @@
-// Grayscale
 $WHITE: #ffffff;
 $BLACK: #000000;
+$GRAY50: #edeff1;
+$GRAY100: #d1d8dc;
+$GRAY300: #94a2ad;
+$GRAY800: #3a474e;
+
+$PRIMARY: #586cf5;
+$SECONDARY: #e5f4fd;
+$WARNING: #ea3a4b;
+
+$BACKGROUND: #f6f7f8;
