+++ src/styles/constants/_levels.scss
@@ -1,1 +1,2 @@
-$TEST: 1
+$LNB: 100;
+$HEADER: 101;
