+++ src/styles/constants/_sizes.scss
@@ -1,2 +1,2 @@
-$LNB_WIDTH: 320px;
-$HEADER_HEIGHT: 80px;
+$HEADER_HEIGHT: 60px;
+$LNB_WIDTH: 300px;
